Demonstrate QRhi-based rendering with a texture and ImGui::Image()
For non-Quick (e.g. the simplewindow example working directly with
QWindow and QRhi) this would be trivial. For the QQuickItem however
extra care is needed due to the threaded rendering model. Implement
the usual pattern for creating a "custom renderer" object that is
then assumed to belong to the Qt Quick render thread and its lifetime
is tied to the QSGNode. This allows correct and safe creation and
destruction of QRhi objects.